Link  DOSKEY article to web archive version
The link for the article on DOSKEY is not available (on 2015-01-15), with a message saying the blog has expired. This replaces it with a version from the web archive. 

An alternative would be to link to the Microsoft information at http://www.microsoft.com/resources/documentation/windows/xp/all/proddocs/en-us/doskey.mspx?mfr=true