Array for dynamic_mappings is incorrect
When you place an array for files with a dynamic_mapping, the following warning appears and no files are found:
Warning: pattern.indexOf is not a function Use --force to continue.

Without the array the files properly expand.  
Note: this could instead be a bug in grunt/task.js with how it looks at obj.extend on line 138.